Fix: neo:deep_fold/3 handling of map values
In the current implementation keys with `map` values are themselves
skipped, under the assumption you are only interested in the leaves.
However that might not always be the case, so this changes that to call
the `fun` on the returning edge.

This is more of a change then fix, as I don't think there's a obvious
correct answer here. It does fix `deep_flatten` as used by
`k4b_debug:repair_created_at` though, and since that is why I wrote this
in the first place, I call this a win.