Fix: update_with
The check for if an key is in a orddict was using the wrong binding.
Also, this fixes a small discrepancy between the badkey error raised by
maps:update_with and neo:update_with.

Taken from kivra_core after the extraction. Would too be hard to
cherry-pick, so copy-paste will have to do.

https://github.com/kivra/kivra_core/commit/9fe1106eb00a5f4f70cd098aa60c5c71422615a6

Co-authored-by: Jimmy Zöger <jimmy.zoger@kivra.com>