Add ExDoc
The `eon` package on hex is not the `eon` package `neo` replaces, which
unfortunately generates a warning.

To actually fix it it needs changes in `eon` itself (publishing its docs
somewhere).

------------------------------------------------------------------------

I tried to ignore it using `skip_undefined_reference_warnings_on`, but
no dice. I guess it's because it generates another type of error.

> warning: documentation references module "eon" but it is limited