Add coverage
Test coverage is always nice, and having it in a machine readable format
allows for more integrations like the VS Code coverage gutter extension.

However, most tools don't understand the coverdata from Erlang (which is really
an ETS table). This adds a conversion script to convert the coverdata to LCOV.