Rewrite part 1: groundwork
This is the beginning of a complete rewrite of neo. Instead of pattern
matching inside each function, this uses separate modules for each data
structure.

The trick is to use records with a module as its tag, allowing dynamic
dispatch. This is similar to Elixir's protocols and structs (which
`neo_reflect` also supports). This also enables third parties to be used
by neo, by implementing a few behaviours.

First is `neo_collection`, representing some sort of container. It's the
base the other behaviours use, and by implementing it you also implement
the others (except less efficiently).

For transversing sequentially use/implement `neo_iterable` and for
transversing in parallel/streaming fashion use/implement `neo_stream`.