Add ETS and persistent terms
This adds neo support for ETS tables and persistent terms, meaning that
you can use `neo_collection` and friends to manipulate them.

This also shows how you can use neo with not just pure data containers,
but also with stateful ones. Sometimes, like for `neo_ets`, the
semantics may differ a bit from something like `dict` or the builtin
maps. The former uses keys `fetch` objects, but its `to_list` returns a
plain list of objects. This is consistent with how `ets` tables work,
but not maps or lists, and that is ok.