Run EUnit tests in inparallel
This runs most tests in parallel, and only those that manipulate global
state (or similar reasons) are run in order.

Unfortunately there's no convenient way to specify test options for just
a subset of tests. What does that mean? If you have `[{application,
neo}, {inparallel, neo_lists}]` EUnit will run the tests in `neo_lists`
twice, first in order and then in parallel.

This is further complicated by rebar running the `rebar.config.script`
before generating the .app file for `neo` (with the list of all its
modules). Instead we just list all the Erlang source files and hope for
the best.

The end result is a major speedup, especially if you have many cores.